+++ HydroPrep/frmHydroPrepRun.cs
@@ -57,7 +57,7 @@ namespace CHaMPWorkbench.HydroPrep
                     {
                         string sInputFile = dbRead.GetString(dbRead.GetOrdinal("InputFile"));
                         if (System.IO.File.Exists(sInputFile))
-                            QueuedRuns.Add(new QueuedRun(dbRead.GetInt32(dbRead.GetOrdinal("ID")), dbRead.GetInt32(dbRead.GetOrdinal("PrimaryVisitID")), sInputFile));
+                            QueuedRuns.Add(new QueuedRun(dbRead.GetInt64(dbRead.GetOrdinal("ID")), dbRead.GetInt64(dbRead.GetOrdinal("PrimaryVisitID")), sInputFile));
                     }
 
                     // Check that at least one run is loaded with a valid input file.
@@ -189,8 +189,8 @@ namespace CHaMPWorkbench.HydroPrep
         /// </summary>
         private class QueuedRun
         {
-            public int BatchRunID { get; internal set; }
-            public int VisitID { get; internal set; }
+            public long BatchRunID { get; internal set; }
+            public long VisitID { get; internal set; }
             public string InputFile { get; internal set; }
 
             /// <summary>
@@ -199,7 +199,7 @@ namespace CHaMPWorkbench.HydroPrep
             /// <param name="nBatchRunID">Workbench Batch run ID</param>
             /// <param name="nVisitID">CHaMP Visit ID</param>
             /// <param name="sInputFile">Model input XML file</param>
-            public QueuedRun(int nBatchRunID, int nVisitID, string sInputFile)
+            public QueuedRun(long nBatchRunID, long nVisitID, string sInputFile)
             {
                 BatchRunID = nBatchRunID;
                 VisitID = nVisitID;
