+++ MainForm.cs
@@ -442,7 +442,7 @@ namespace CHaMPWorkbench
                             mnuQuery.Tag = new UserQueries.frmQueryProperties.UserQueryTag(
                                 DBCon.ConnectionString,
                                 dbRead.GetString(dbRead.GetOrdinal("QueryText")),
-                                dbRead.GetInt32(dbRead.GetOrdinal("QueryID")),
+                                dbRead.GetInt64(dbRead.GetOrdinal("QueryID")),
                                 dbRead.GetString(dbRead.GetOrdinal("Title")));
 
                             mnuQuery.Click += UserQueries.frmQueryProperties.RunUserQuery;
@@ -1734,7 +1734,7 @@ namespace CHaMPWorkbench
                         ToolStripMenuItem mnuQuery = new ToolStripMenuItem(dbRead.GetString(dbRead.GetOrdinal("Title")));
 
                         // Build a tag that contains everything the query needs to run
-                        mnuQuery.Tag = new naru.db.NamedObject(dbRead.GetInt32(dbRead.GetOrdinal("ItemID")), dbRead.GetString(dbRead.GetOrdinal("Title")));
+                        mnuQuery.Tag = new naru.db.NamedObject(dbRead.GetInt64(dbRead.GetOrdinal("ItemID")), dbRead.GetString(dbRead.GetOrdinal("Title")));
                         mnuQuery.Click += this.ShowMetricReviewForm;
                         metricReviewToolStripMenuItem.DropDownItems.Add(mnuQuery);
                     }
